requirements.txt
streamlit
crewai
crewai-tools
langchain-openai
pypdf
sentence-transformers
pinecone
langchain-pinecone
groundx